requirements.txt
# ── Core ML ───────────────────────────────────────────────────────────────────
torch>=2.2.0
torchvision>=0.17.0
Pillow>=10.2.0
numpy>=1.26.0

# ── Inference (edge) ──────────────────────────────────────────────────────────
onnx>=1.15.0               # required for torch.onnx.export
onnxruntime>=1.17.0        # CPU inference; use onnxruntime-gpu for GPU edge

# ── API ───────────────────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9    # required for UploadFile support
pydantic>=2.6.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0              # async test client for FastAPI
